Added Island controller, repository and model and OneToMany relationship with worlds

+++ src/main/java/com/sdca/api/repository/WorldRepository.java
@@ -3,5 +3,5 @@ package com.sdca.api.repository;
 import com.sdca.api.model.World;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WorldRepository extends CrudRepository<World, Integer> {
+public interface WorldRepository extends CrudRepository<World, Long> {
 }
